Extract generic ThumbnailRenderer from PatternThumbnail
The offscreen render (bounding box, RenderTexture, read-back, scale-to-square,
in-memory cache) had nothing pattern-specific about it. Move it into a content-
agnostic ThumbnailRenderer that takes a back-to-front list of positioned sprites;
PatternThumbnail now just builds the variant's floor/object/roof sprites and hands
them over. Behavior is unchanged, and the map-loader thumbnails can reuse the same
renderer (plus the shared buildTileSprite/buildSpriteObject helpers).
