+++ src/pattern/PatternThumbnail.cpp
@@ -1,25 +1,19 @@
 #include "pattern/PatternThumbnail.h"
 
-#include <algorithm>
-#include <limits>
 #include <memory>
-#include <optional>
 #include <vector>
 
-#include <SFML/Graphics.hpp>
-#include <spdlog/spdlog.h>
+#include <SFML/Graphics/Sprite.hpp>
 
 #include <QDateTime>
 #include <QFileInfo>
-#include <QImage>
-#include <QPainter>
-#include <QPixmapCache>
 
 #include "editor/HexagonGrid.h"
 #include "editor/Object.h"
 #include "pattern/Pattern.h"
 #include "pattern/PatternSprite.h"
 #include "pattern/PatternStamper.h"
+#include "ui/rendering/ThumbnailRenderer.h"
 
 namespace geck::pattern {
 
@@ -27,18 +21,10 @@ namespace {
 
     // In-memory cache key, invalidated by the source file's mtime. Only a handful of
     // small thumbnails are live at a time, so they are rendered on demand and kept in
-    // Qt's global pixmap cache rather than written to disk.
+    // the shared in-memory cache rather than written to disk.
     QString cacheKey(const QString& sourcePath, int size) {
         const qint64 mtime = QFileInfo(sourcePath).lastModified().toSecsSinceEpoch();
         return sourcePath + '|' + QString::number(mtime) + '|' + QString::number(size);
-    }
-
-    // Copy an SFML RGBA image into an owned QImage.
-    QImage toQImage(const sf::Image& image) {
-        const sf::Vector2u sz = image.getSize();
-        return QImage(image.getPixelsPtr(), static_cast<int>(sz.x), static_cast<int>(sz.y),
-            QImage::Format_RGBA8888)
-            .copy();
     }
 
 } // namespace
@@ -46,11 +32,8 @@ namespace {
 QPixmap PatternThumbnail::forPattern(const Pattern& pattern, const QString& sourcePath,
     resource::GameResources& resources, const HexagonGrid& hexgrid, int size) {
     const QString key = sourcePath.isEmpty() ? QString() : cacheKey(sourcePath, size);
-    if (!key.isEmpty()) {
-        QPixmap cached;
-        if (QPixmapCache::find(key, &cached)) {
-            return cached;
-        }
+    if (auto hit = ThumbnailRenderer::cached(key)) {
+        return *hit;
     }
 
     if (pattern.variants.empty()) {
@@ -58,7 +41,7 @@ QPixmap PatternThumbnail::forPattern(const Pattern& pattern, const QString& sour
     }
 
     // Build positioned sprites for variant 0 at its authored positions: floor and roof
-    // tiles plus objects.
+    // tiles plus objects. They are kept alive here while ThumbnailRenderer draws them.
     const PatternVariant& variant = pattern.variants.front();
     const PatternStamper::Plan plan = PatternStamper::plan(variant, variant.anchorHex);
 
@@ -77,72 +60,20 @@ QPixmap PatternThumbnail::forPattern(const Pattern& pattern, const QString& sour
         }
     }
 
-    if (floorSprites.empty() && roofSprites.empty() && objects.empty()) {
-        return {}; // no resolvable art
-    }
-
-    // Union of all sprite bounds, with a small margin.
-    float minX = std::numeric_limits<float>::max();
-    float minY = std::numeric_limits<float>::max();
-    float maxX = std::numeric_limits<float>::lowest();
-    float maxY = std::numeric_limits<float>::lowest();
-    const auto extend = [&](const sf::FloatRect& b) {
-        minX = std::min(minX, b.position.x);
-        minY = std::min(minY, b.position.y);
-        maxX = std::max(maxX, b.position.x + b.size.x);
-        maxY = std::max(maxY, b.position.y + b.size.y);
-    };
-    for (const auto& sprite : floorSprites) {
-        extend(sprite.getGlobalBounds());
+    // Flatten into the editor's draw order: floor tiles, then objects, then roof tiles.
+    std::vector<const sf::Sprite*> ordered;
+    ordered.reserve(floorSprites.size() + objects.size() + roofSprites.size());
+    for (const sf::Sprite& sprite : floorSprites) {
+        ordered.push_back(&sprite);
     }
     for (const auto& object : objects) {
-        extend(object->getSprite().getGlobalBounds());
+        ordered.push_back(&object->getSprite());
     }
-    for (const auto& sprite : roofSprites) {
-        extend(sprite.getGlobalBounds());
+    for (const sf::Sprite& sprite : roofSprites) {
+        ordered.push_back(&sprite);
     }
-    constexpr float pad = 4.0f;
-    minX -= pad;
-    minY -= pad;
-    maxX += pad;
-    maxY += pad;
-    const auto width = static_cast<unsigned int>(std::max(1.0f, maxX - minX));
-    const auto height = static_cast<unsigned int>(std::max(1.0f, maxY - minY));
 
-    sf::RenderTexture renderTexture;
-    if (!renderTexture.resize({ width, height })) {
-        spdlog::warn("PatternThumbnail: failed to create {}x{} render texture", width, height);
-        return {};
-    }
-    renderTexture.setView(sf::View(sf::FloatRect({ minX, minY },
-        { static_cast<float>(width), static_cast<float>(height) })));
-    renderTexture.clear(sf::Color::Transparent);
-    // Draw order matches the editor: floor tiles, then objects, then roof tiles on top.
-    for (const auto& sprite : floorSprites) {
-        renderTexture.draw(sprite);
-    }
-    for (const auto& object : objects) {
-        renderTexture.draw(object->getSprite());
-    }
-    for (const auto& sprite : roofSprites) {
-        renderTexture.draw(sprite);
-    }
-    renderTexture.display();
-
-    const QImage rendered = toQImage(renderTexture.getTexture().copyToImage());
-    const QImage scaled = rendered.scaled(size, size, Qt::KeepAspectRatio, Qt::SmoothTransformation);
-
-    QImage canvas(size, size, QImage::Format_RGBA8888);
-    canvas.fill(Qt::transparent);
-    QPainter painter(&canvas);
-    painter.drawImage((size - scaled.width()) / 2, (size - scaled.height()) / 2, scaled);
-    painter.end();
-
-    QPixmap pixmap = QPixmap::fromImage(canvas);
-    if (!key.isEmpty()) {
-        QPixmapCache::insert(key, pixmap);
-    }
-    return pixmap;
+    return ThumbnailRenderer::render(ordered, size, key);
 }
 
 } // namespace geck::pattern
